pyproject.toml
[tool.poetry]
name = "drf-project"
version = "0.1.0"
description = ""
authors = ["Said Magomedov <saidmagomedov037@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.3"
python-dotenv = "^1.1.1"
djangorestframework = "^3.16.0"
psycopg2 = "^2.9.10"
pillow = "^11.2.1"
black = "^25.1.0"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
